Package References

Package References

DemoCards.cardtheme — Function.

cardtheme(theme::AbstractString) -> path

Currently supported themes are:

minimal

source

DemoCards.makedemos — Method.

makedemos(source::String) -> path

Make a demo page file and return the path to it.

source is the root path to the demos folder, by default it's relative path to docs.

Processing pipeline:

analyze the folder structure source and loading all available configs.

copy assets

preprocess demo files and save it

save/copy cover images

Note

By default, the source demo files are read, processed and save to docs/src/democards, so if you put all source demo files in docs/src, there will be a duplication of files and assets.

Keywords

root::String: root path to the whole documentaion. By default docs.

destination::String: By default democards.

Examples

You only need to call this function before Documenter.makedocs, and pass the result to it.

format = Documenter.HTML(edit_branch = "master",
                         assets = [joinpath("assets", "style.css")])

examples = makedemos("examples")

makedocs(format = format,
         pages = [
            "Home" => "index.md",
            "Examples" => examples,
         ])

Warning

Currently, there's no guarantee that this function works for unconventional documentation folder structure. By convention, it is:

.
├── Project.toml
├── docs
│   ├── make.jl
│   └── src
├── src
└── test

source

DemoCards.DemoPage — Type.

struct DemoPage <: Any
DemoPage(root::String)

Constructs a demo page object.

Fields

Besides the root path to the demo page folder root, this struct has some other fields:

title: page title

template: template content of the demo page.

sections: demo sections found in root

Configuration

You can manage an extra config.json file to customize rendering of a demo page. Supported items are:

order: specify the sections order. By default, it's case-insensitive alphabetic order.

template: path to template filename. The content of the template file should has one and only one {{{democards}}}.

title: specify the title of this demo page. By default, it's the folder name of root. Will be override by template.

The following is an example of config.json:

{
    "template": "template.md",
    "order": [
        "basic",
        "advanced"
    ]
}

Examples

The following is the simplest folder structure of a DemoPage:

demos
└── basic
    ├── demo_1.md
    ├── demo_2.md
    ├── demo_3.md
    ├── demo_4.md
    ├── demo_5.md
    ├── demo_6.md
    ├── demo_7.md
    └── demo_8.md

Note

A DemoPage doesn't manage demo files directly, so here you'll need a DemoSection basic to manage them.

The following is a typical folder structure of a DemoPage:

demos
├── advanced
│   ├── advanced_demo_1.md
│   └── advanced_demo_2.md
├── basic
│   ├── part1
│   │   ├── basic_demo_1_1.md
│   │   └── basic_demo_1_2.md
│   └── part2
│       ├── config.json
│       ├── basic_demo_2_1.md
│       └── basic_demo_2_2.md
├── config.json
└── template.md

Warning

A section should only hold either subsections or demo files. A folder that has both subfolders and demo files (e.g., *.md) is invalid.

See also: MarkdownDemoCard, DemoSection

source

DemoCards.DemoSection — Type.

struct DemoSection <: Any
DemoSection(root::String)

Constructs a demo section that holds either nested subsections or demo cards.

Fields

Besides the root path to the demo section folder root, this struct has some other fields:

cards: demo cards found in root

subsections: nested subsections found in root

Configuration

You can manage an extra config.json file to customize rendering of a demo section. Supported items are:

order: specify the cards order or subsections order. By default, it's case-insensitive alphabetic order.

title: specify the title of this demo section. By default, it's the folder name of root.

The following is an example of config.json:

{
    "title": "learn by examples"
    "order": [
        "quickstart.md",
        "array.md"
    ]
}

Warning

You can't specify files or foldernames in other folders.

Examples

The following is the simplest folder structure of a DemoSection:

section
├── demo_1.md
├── demo_2.md
├── demo_3.md
├── demo_4.md
├── demo_5.md
├── demo_6.md
├── demo_7.md
└── demo_8.md

The following is a typical folder structure of a DemoSection:

section
├── config.json
├── part1
│   ├── demo_21.md
│   └── demo_22.md
└── part2
    ├── config.json
    ├── demo_23.md
    └── demo_24.md

Warning

A section should only hold either subsections or demo files. A folder that has both subfolders and demo files (e.g., *.md) is invalid.

See also: MarkdownDemoCard, DemoPage

source

DemoCards.MarkdownDemoCard — Type.

struct MarkdownDemoCard <: AbstractDemoCard
MarkdownDemoCard(path::String)

Constructs a markdown-format demo card from existing markdown file path.

Fields

Besides path, this struct has some other fields:

path: path to the source markdown file

cover: path to the cover image

id: cross-reference id

title: one-line description of the demo card

description: multi-line description of the demo card

Configuration

You can pass additional information by adding a YAML front matter to the markdown file. Supported items are:

cover: relative path to the cover image. If not specified, it will use the first available image link, or all-white image if there's no image links.

description: a multi-line description to this file, will be displayed when the demo card is hovered. By default it uses title.

id: specify the id tag for cross-references. By default it's infered from the filename, e.g., simple_demo from simple demo.md.

title: one-line description to this file, will be displayed under the cover image. By default, it's the name of the file (without extension).

An example of the front matter:

---
title: passing extra information
cover: cover.png
id: non_ambiguious_id
description: this demo shows how you can pass extra demo information to DemoCards package.
---

See also: DemoSection, DemoPage

source

DemoCards.democard — Method.

democard(path::String)::T

Constructs a concrete AbstractDemoCard instance.

The return type T is determined by the extension of the path to your demofile. Currently supported types are:

MarkdownDemoCard

source

DemoCards.get_default_order — Method.

return case-insensitive alphabetic order

source

DemoCards.load_title — Method.

load_title(page::DemoPage, config)

load title from config using the following priority rules:

parse the title out from the template file

use config["title"]

fallback: use the folder name

source

DemoCards.parse — Method.

parse(card::MarkdownDemoCard)

parse out configuration of markdown files and return it as a Dict.

Possible configuration resources are:

YAML front matter

image links

Note

Users of this function need to use haskey to check if keys are existed. They also need to validate the values.

source

DemoCards.parse_markdown — Method.

parse_markdown(path::String)

parse the template file of page and return a configuration dict.

Currently supported items are: title, id.

source

DemoCards.save_cover — Method.

save_cover(path::String, card::AbstractDemoCard)

process the cover image and save it.

source

DemoCards.save_markdown — Method.

save_markdown(root::String, card::MarkdownDemoCard)

process the original markdown file and save it.

The processing pipeline is:

strip the front matter

insert a level-1 title and id

source